backend/requirements.txt
annotated-types==0.6.0
anyio==4.3.0
beautifulsoup4==4.12.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
distro==1.9.0
exceptiongroup==1.2.1
fastapi==0.110.2
h11==0.14.0
html5lib==1.1
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
jsonpath-rw==1.4.0
lxml==5.2.1
metar==1.11.0
openai==1.14.2
playsound==1.3.0
ply==3.11
pydantic==2.7.0
pydantic_core==2.18.1
pyflightdata==0.8.6.2
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
starlette==0.37.2
tqdm==4.66.2
typing_extensions==4.11.0
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
webencodings==0.5.1
websockets==12.0
